refactor(operations): 抽取共享操作表单
